Add files via upload
The proposed system is to develop a ensemble learning model which
combines the prediction of other deep learning models. In this project different
deep learning models such as CNN, VGG and ResNet. When these deep
learning model predictions are combined together it gives higher accuracy
than the accuracy produced by a single deep learning model. First the data
is collected FER2013 dataset is used in this project. The collected data is
preprocessed. Pre-processing includes resizing, augmentation and applying
gaussian and laplacian filter. These filters can be used to find the edge of the
face. Then the data is trained using the deep learning models and finally the
accuracy is calculated. Finally the stack ensemble is build using the predictions
of CNN models.

The objective of the project is to develop a robust and accurate
model that can accurately classify the emotional states depicted in images. This
involves training a deep learning model, such as a CNN and ResNet, using the
FER dataset, which consists of labeled facial expressions representing various
emotions. This project also includes building an stack ensemble model and
provide good accuracy in classification. For ensemble learning random forest is
used as the meta model